finished exercise jsx components and props
